requirements.txt
requests
ccxt
pandas
plotly
openai>=1.12.0
streamlit>=1.30.0
pypdf
pdf2image
pytesseract
Pillow

google-generativeai
PyPDF2